+++ R/fastpos.R
@@ -28,7 +28,8 @@ NULL
 #' stop execution ourselves (and also return nothing).
 #' @noRd
 stop_quietly <- function() {
-  blankMsg <- sprintf("\r%s\r", paste(rep("", getOption("width")-1L), collapse=" "));
+  blankMsg <- sprintf("\r%s\r", paste(rep("", getOption("width")-1L),
+                                      collapse=" "));
   stop(simpleError(blankMsg));
 }
 
@@ -74,16 +75,7 @@ create_pop <- function(rho, size) {
 #' Run simulation for one specific correlation.
 #'
 #' @param rho Population correlation.
-#' @param pop_size Population size.
-#' @param sample_size_max Number of participants for each study.
-#' @param n_studies Number of studies to run.
-#' @param precision Precision around the correlation which is acceptable
-#' @param precision_rel Whether the precision is absolute (rho+-precision or
-#'   relative rho+-rho*precision).
-#' @param sample_size_min Minimum sample size to start in corridor of stability.
-#' @param confidence_levels Confidence levels for point of stability. This
-#'   corresponds to the quantile of the distribution of all found critical
-#'   sample sizes. Defaults to c(.8, .9, .95).
+#' @inheritParams find_critical_pos
 #' @return A list with two elements, (1) a data frame called "summary"
 #'   containing all the above information as well as the critical sample sizes
 #'   (points of stability) for the confidence-levels specified and (2) vector
@@ -93,10 +85,12 @@ create_pop <- function(rho, size) {
 #' find_one_critical_pos(rho = 0.5)
 #' @noRd
 find_one_critical_pos <- function(rho, sample_size_min = 20,
-                                  sample_size_max = 1000, n_studies = 1000,
+                                  sample_size_max = 1000,
+                                  replace = TRUE, n_studies = 1000,
                                   pop_size = 1e6, precision = .1,
                                   precision_rel = T,
-                                  confidence_levels = c(.8, .9, .95)) {
+                                  confidence_levels = c(.8, .9, .95),
+                                  n_cores = 1) {
   # create corridor of stability
   if (precision_rel) {
     limits <- rho * (1 + c(-1, 1) * precision)
@@ -113,8 +107,27 @@ find_one_critical_pos <- function(rho, sample_size_min = 20,
   rho_pop <- stats::cor(x, y)
 
   # create dist of pos
-  res <- simulate_pos(x, y, n_studies, sample_size_min, sample_size_max, T,
-                       lower_limit, upper_limit)
+
+  # we will use 1 normal invocation and n_cores - 1 futures, this way we have
+  # a progress bar
+  if (n_cores > 1){
+    #future::plan(multisession)
+    f <- list()
+    for (ii in seq(n_cores - 1)) {
+      f[[ii]] <- future::future({
+        simulate_pos(x, y, ceiling(n_studies/(n_cores)), sample_size_min,
+                     sample_size_max, replace, lower_limit, upper_limit)
+      })
+    }
+    res <- simulate_pos(x, y, ceiling(n_studies/n_cores), sample_size_min,
+                        sample_size_max, replace, lower_limit, upper_limit)
+    v <- unlist(lapply(f, FUN = future::value))
+    res <- c(res, v)
+  } else {
+    res <- simulate_pos(x, y, n_studies, sample_size_min, sample_size_max, T,
+                        lower_limit, upper_limit)
+  }
+
   # on interruption, C++ will return -1 (if R interrupts by itself, nothing
   # will be returned, it just stops)
   if (length(res) == 1) {
@@ -132,7 +145,7 @@ find_one_critical_pos <- function(rho, sample_size_min = 20,
                           sample_size_max = sample_size_max,
                           lower_limit=lower_limit,
                           upper_limit=upper_limit,
-                          n_studies = n_studies,
+                          n_studies = length(res),
                           n_not_breached = n_not_breached),
               n = res
               ))
@@ -156,10 +169,12 @@ find_one_critical_pos <- function(rho, sample_size_min = 20,
 #' @param n_studies Number of studies to run for each rho (defaults to 10e3).
 #' @param sample_size_min Minimum sample size for each study (defaults to 20).
 #' @param sample_size_max Maximum sample size for each study (defaults to 1e3).
+#' @param replace Whether drawing samples is with replacement or not.
 #' @param confidence_levels Confidence levels for point of stability. This
 #'   corresponds to the quantile of the distribution of all found critical
 #'   sample sizes (defaults to c(.8, .9, .95)).
 #' @param pop_size Population size (defaults to 1e6).
+#' @param n_cores Number of cores to use for simulation.
 #' @return A data frame containing all the above information, as well as the
 #'   points of stability.
 #' @examples
@@ -168,15 +183,20 @@ find_one_critical_pos <- function(rho, sample_size_min = 20,
 #' @export
 find_critical_pos <- function(rhos, precision = 0.1, precision_rel = FALSE,
                               sample_size_min = 20, sample_size_max = 1000,
+                              replace = TRUE,
                               n_studies = 10000,
                               confidence_levels = c(.8, .9, .95),
-                              pop_size = 1e6) {
+                              pop_size = 1e6,
+                              n_cores = 1) {
+  defaultplan <- options("future.plan")[[1]]
+  options("future.plan" = "multisession")
   result <- mapply(find_one_critical_pos, rhos,
                    sample_size_max = sample_size_max,
                    sample_size_min = sample_size_min,
                    n_studies = n_studies,
                    precision = precision,
                    precision_rel = precision_rel,
+                   n_cores = n_cores,
                    MoreArgs = list(confidence_levels = confidence_levels),
                    SIMPLIFY = F)
   summary <- lapply(result, function(x) x[[1]])
@@ -188,6 +208,7 @@ find_critical_pos <- function(rhos, precision = 0.1, precision_rel = FALSE,
             stability", ".\nIncrease sample_size_max and rerun the simulation.",
             sep = "")
   }
+  options("future.plan" = defaultplan)
   summary
 }