+++ inst/doc/fastpos.R
@@ -1,34 +1,23 @@
 ## ---- include = FALSE---------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>"
+  comment = "#>",
+  cache = TRUE
 )
 
 ## ----setup--------------------------------------------------------------------
 library(fastpos)
 set.seed(19950521)
 
-## -----------------------------------------------------------------------------
-find_critical_pos(rho = .7, sample_size_min = 20, sample_size_max = 1000,
-                  n_studies = 10000)
+## ----message=TRUE, warning=TRUE, paged.print=TRUE-----------------------------
+find_critical_pos(rhos = seq(.1, .7, .1), sample_size_max = 1e3,
+                  n_studies = 10e3)
 
-## -----------------------------------------------------------------------------
-find_critical_pos(rho = .7, sample_size_min = 20, sample_size_max = 400,
-                  n_studies = 10000)
-
-## -----------------------------------------------------------------------------
-find_critical_pos(rho = .7, sample_size_min = 20, sample_size_max = 1000,
-                  n_studies = 10000, confidence_levels = c(.6, .85))
-
-## -----------------------------------------------------------------------------
-find_critical_pos(rho = c(.5, .7), sample_size_min = 20, sample_size_max = 2500,
-                  n_studies = 10000, precision = .10, precision_rel = T)
-
-## ---- fig.width = 6, fig.height = 4.2-----------------------------------------
+## ----fig.height=4.8, fig.width=6.4--------------------------------------------
 pop <- create_pop(0.5, 1000000)
 pos <- simulate_pos(x_pop = pop[,1],
                     y_pop = pop[,2],
-                    n_studies = 1000,
+                    n_studies = 10000,
                     sample_size_min = 20,
                     sample_size_max = 1000,
                     replace = T,
@@ -38,7 +27,26 @@ hist(pos, xlim = c(0, 1000), xlab = c("Point of stability"),
      main = "Histogram of points of stability for rho = .5+-.1")
 quantile(pos, c(.8, .9, .95), na.rm = T)
 
-## ----message=TRUE, warning=TRUE, paged.print=TRUE-----------------------------
-find_critical_pos(rho = seq(.1, .7, .1), sample_size_max = 1000,
-                  n_studies = 10000)
+## ----parallel1, message=FALSE, warning=FALSE----------------------------------
+onecore <- function() {find_critical_pos(0.5)}
+multicore <- function() {find_critical_pos(0.5, n_cores = future::availableCores())}
+microbenchmark::microbenchmark(onecore(), multicore(), times = 10)
+
+## ----parallel2, message=FALSE, warning=FALSE----------------------------------
+onecore <- function() {find_critical_pos(0.5, n_studies = 1e5)}
+multicore <- function() {find_critical_pos(0.5, n_studies = 1e5,
+                                           n_cores = future::availableCores())}
+microbenchmark::microbenchmark(onecore(), multicore(),
+                               times = 10)
+
+## ----eval=FALSE, include=FALSE------------------------------------------------
+#  speedup <- function(n_max, n_min){
+#    (n_max*(n_max+1)/2-n_min*(n_min-1)/2)/(2*n_max-n_min)
+#  }
+#  speedup(1000, 20)
+#  speedup2 <- function(n_max, n_min, cpos){
+#    (n_max*(n_max+1)/2-n_min*(n_min-1)/2)/(2*n_max-cpos)
+#  }
+#  speedup2(1000, 20, 119)
+#  # get the 50% quantile to estimate how long on average it takes for fastpos to stop.